Fix dhcp issues and improve public API
- Use qconnectivity daemon for dhcp requests. Since dhcp requests are
executed in other process we don't block gui thread for this lengthy
operation.

- Why not to use "do_dhcp_request" - it is a legacy implementation of a
dhcp client and is not used anywhere in the Android source code itself.
It appears that do_dhcp_request was not removed from the libhardware_legacy
to keep compatibility for others, whose code might still depend on it.

- Differnet changes to the internal implementation and the public API.

- Add 'Interface' singleton type, installing a singleton type allows
developers to provide arbitrary functionality to a client without requiring
individual instances of the type to be instantiated by the client. We use
this to determine if Android has wifi interface.

Change-Id: I836f3a2a587b1ebf9f670ed08b10fe3483504b9e
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@digia.com>