Improve QDroidUtils IP address retrieval
Instead of returning the first IP address from eth0, return
a comma-separated list. This ensures that the IPv4 address
is shown in Launcher Settings demo, important as it's
required for setting up the connection to Raspberry Pi in
Qt Creator (sometimes, IPv6 address is shown as first
entry).

Also, cleans up the garbage (trailing %eth0) returned for
IPv6 addresses.

Change-Id: Ibced367ba443ac76c2ad2a4f2589c9edcbc74b7f
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@digia.com>