Remove unnecessary roles from QWifiNetworkList
Currently there are 2 ways to access the same property from list
view's delegate:

network.ssid
ssid

This is not needed, from the API point of view it is more readible
when properties are accesed with network.* prefix. As is done in
launchersettings demo.

"network" role returns QWifiNetwork object which exposes the required
properties with Q_PROPERTY.

Change-Id: Idd55891d44c4f356a6774575508a38729c60c335
Reviewed-by: Andy Nichols <andy.nichols@digia.com>