Rename misleading class name
The API is:

ListView {
   model: wifiManager.networks
}

currently if user calls print(wifiManager.networks) it will print:

QWifiNetworkList

The docs of ListView states that model property expects a subclass
of QAbstractListModel. QWifiNetworkList sounds like a subclass of
QList which is misleading.

This patch renames the class to QWifiNetworkListModel. This does not
change the public API (doesn't brake compatibility in any way), just
improves the documentation of library.

Task-number: QTEE-445

Change-Id: I71f07cb22617f3a7f41a0403d1c63c4357310325
Reviewed-by: Andy Nichols <andy.nichols@digia.com>