Doc: Add Known Issues page
Restore Known Issues once documented on qt.digia.com as part of
the documentation.

Move Troubleshooting items that are clear known issues to correct
page.

For each item, replace \section titles with just a bold text to
get rid of the overly-complex Table of Contents. Instead of the
TOC, use a custom table in the beginning of the table.

Task-number: QTEE-799
Change-Id: Id53775146f1084cd59891527badcc7aaeea51534
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@theqtcompany.com>