B2Qt utils: implement backlight setting for eLinux
This at least works for Sabre and BD-SL, and should be easy enough to
extend if other boards have the controlling device in a different
path.

In case of multiple screens, this version just sets all of them. This
is sufficient for demos and the common single-screen case. Next
version will add API to address each screen separately.

Change-Id: I1bbd5201474bd92d20efba6d55b11339200c141a
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>