Pass "-D nl80211,wext" to supplicant by default
nl80211 is the current standard, but not all wireless chip's
modules support it, wext is currently deprecated, but still
widely supported. Supplicant will choose which of the these
kernel interfaces to use for controlling a wireless driver on
a specific device.

This for example fixes the issue of QtWifi not working properly
on a Nitrogen6_Max board with a wifi chip that comes on it by
default.

Change-Id: I6f24df7126e761bf196ec6ee76918a96d03307a5
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>