Fix signal calculation for WEXT-based drivers
This patch introduces a behavior change. Previously the
'SignalStrength' role was interpreted as dBm. This was working
fine for nl80211 based drivers, for WEXT-based drivers it
contained garbage values. Now the 'SignalStrength' role will
always be interpreted as percentage.

The wpa_supplicant passes back only the 'level' of the scanned
BSS, which with nl80211-based driver is almost always dBm (val < 0).
WEXT-based driver has its own interpretation of 'level' - it
can contain dBm, percentage [0-100] (and some other values for
historical reasons, but these are converted by the wpa_supplicant
to dBm). This is what I have concluded from looking at NetworkManager
and wpa_supplicant sources and git history. Even the official GUI tool
wpa_gui for the wpa_supplicant is out-of-date and can not be used as
a referance.

Change-Id: I70d15e26e0a1700ad0f7d19d23cbade2d3d46054
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@theqtcompany.com>