Doc: Add a custom note on the online landing page
The installer now opens a browser with the latest online docs. As
the version the user is installing may not be the very latest one,
add a note on the landing page reminding that version-specific
documentation is available in Qt Creator Help.

This note is shown only when launching from the installer - i.e,
the URL has the 'welcome' query string:

    http://doc.qt.io/QtForDeviceCreation/index.html?welcome

Task-number: QTEE-964
Change-Id: Iaf373d97c960d36d206f93ec7d5b61086adda86c
Reviewed-by: Samuli Piippo <samuli.piippo@theqtcompany.com>