Doc: Remove generic Device Creation documentation
... and replace the documentation configuration with a one specific
to Boot to Qt Utils. The removed documentation is available in a
dedicated repository, tqtc-boot2qt/doc.

The name of this module in 5.6 branch will be 'Qt for Device
Creation: Add-On Modules'.

The doc namespace and config files already reflect the name
'Qt Device Utilities', but that is not visible to the user in
5.6 (but will be in 5.7).

Change-Id: Ib5a98d57b6ff5e7c29901f42697266f396abad82
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>