Conan: Use "scm" revision mode when exporting from .git repository
We want to match the commit ID along with the
RREV (recipe revision). This way we can pin exact Conan packages
to known sha1/tag which makes releasing easier.

https://docs.conan.io/en/latest/reference/conanfile/attributes.html#revision-mode

We have two use cases where the conan export is being executed:
  - Conan export from src tar ball
    - .git directory is not present
  - Coin/CI: conan export from git repository

The "scm" feature will not work if the ".git" repository does not
exist. If this is the case then use the default "hash" mode for the
recipe revision.

Task-number: QTQAINFRA-4325
Change-Id: I5e3ac718333cd47b95d5a8886b33b1b5b7b31300
Reviewed-by: Toni Saario <toni.saario@qt.io>