Update commercial license headers
Updated header.COMM to the files in tqtc-qtdeviceutilities.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4943
Change-Id: Iffbbfc261d34161e769961ea4f4711be1cf1c0a7
Reviewed-by: Pasi Petäjäjärvi <pasi.petajajarvi@qt.io>