Update commercial license headers
Updated header.COMM to the files in tqtc-qtdeviceutilities.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-5900
Change-Id: I3482af415eb91eeb66155483b1f1b79252232e04
Reviewed-by: Pasi Petäjäjärvi <pasi.petajajarvi@qt.io>